backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery
redis
cloudinary
ably
email-validator
python-dotenv
cors
fastapi-cors
pytest
pytest-asyncio
httpx
pytesseract
Pillow
stytch
reportlab
